+++ sensor/ina219_driver.h
@@ -25,8 +25,8 @@
 typedef struct {
     // I2C address of the sensor
     uint8_t i2c_address;
-    // Shunt resistor in micro ohms
-    uint32_t shunt_resistor;
+    // Shunt resistor in ohms
+    double shunt_resistor;
 } Ina219Config;
 
 SensorDriver* ina219_driver_alloc(const Ina219Config* config);
